Youtube routes externalized to /routes/ytmusic.js
